app/pyproject.toml
[project]
name = "mai_qrcode"
version = "1.0.0"
description = "Maimai QR Code Client for Wechat"
authors = [
    {name = "FurryR", email = "awathefox@gmail.com"},
]
dependencies = ["pyautogui>=0.9.54", "aiohttp>=3.12.15", "fastapi>=0.116.1", "uvicorn>=0.35.0", "python-xlib>=0.33", "dotenv>=0.9.9", "asyncio>=4.0.0", "mss>=10.0.0", "opencv-python>=4.12.0.88", "pyzbar>=0.1.9", "numpy>=2.2.6"]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
